Scope particle network to Hero section only
Move canvas/particle animation from the global AmbientBackground into a
new HeroParticles client component. The effect is now contained within
the hero section (clipped by overflow-hidden), leaving all other sections
clean. AmbientBackground retains only the solid background and film grain.
